Fixed: Mask loading in higher res tiles
The low res tiles are presented in the threejs overlay first and the higher resolution tiles and layers are loaded in later.  As the textures are updated the tiles that now have a texture file are then rendered visible and it hides the loading in a similar way to openseadragon.