Fixed: Moving light resets intensity sliders
I have changed the way certain props are updated by making them control specific and have changed the way the reset function works.  This fix not only fixes the light sliders resetting on direction change, it also allows for position and intensity to be remembered in between relight toggling until reset is pressed.

Closes: #78