Fixed: new lighting mode
This adds in more light control, the light is now more exaggerated and there is a helper to be able to visualise directional light direction in the scene.  The directional light now follows the intersection and the light target is positioned in the centre of whatever you are looking at.