Fixed: Shift layers is working
May need some cleaning up of variable names in future, but this change works quite well.  We take full control of the layers state by populating it early with the choices order and using the redux state of the application to manage ordering of the layers and if the user uses the native features of Mirador to move around the layers, we respect that order.