WIP: custom saga events
Putting this in a branch because it's valuable code.  However, going to remove saga changes as it impacts performance for not much gained.  For now we will roll back some changes, and just make the sidebar open by default on larger screens on load, but turned off on smaller screens