New: Add help button with dialog
In order to provide help text we need to open a dialog box and close it using another button.
We can add the help content to this including HTML text and images to guide a user through
how to use the plug-in in more detail.  I have also tweaked the alignment of the controls to look better in mobile view with this change.