WIP: Experimenting with alternative ways to move light around
Tried using transform controls in threejs, this doesn't work well with Mirador rotate, so instead I have a partially implemented dragable controller in the plug-in interface instead.  It's partially implemented because the threejs transform needs tearing out and the props from moving around the Mirador plug-in control need adding in instead.